refactor: one class per file, move ErrorResponse to shared actions package
Addresses PR #31 review feedback:
- Split PostResponse.kt into individual files per DTO
- Move ErrorResponse to adapters/misk/actions/ (not posts-specific)

ref: #13
Amp-Thread-ID: https://ampcode.com/threads/T-019c5920-739b-77ef-92f9-d3f37d048d9e
