+++ src/main/kotlin/xyz/block/buildersyndicate/adapters/misk/actions/ErrorResponse.kt
@@ -1,0 +1,6 @@
+package xyz.block.buildersyndicate.adapters.misk.actions
+
+data class ErrorResponse(
+  val error: String,
+  val message: String,
+)
