+++ src/main/kotlin/xyz/block/buildersyndicate/adapters/misk/actions/posts/CreatePostRequest.kt
@@ -1,0 +1,6 @@
+package xyz.block.buildersyndicate.adapters.misk.actions.posts
+
+data class CreatePostRequest(
+  val title: String,
+  val body: String,
+)
