+++ src/main/kotlin/xyz/block/buildersyndicate/adapters/misk/actions/posts/PostListResponse.kt
@@ -1,0 +1,5 @@
+package xyz.block.buildersyndicate.adapters.misk.actions.posts
+
+data class PostListResponse(
+  val posts: List<PostResponse>,
+)
