+++ src/main/kotlin/xyz/block/buildersyndicate/adapters/misk/actions/posts/UpdatePostRequest.kt
@@ -1,0 +1,6 @@
+package xyz.block.buildersyndicate.adapters.misk.actions.posts
+
+data class UpdatePostRequest(
+  val title: String,
+  val body: String,
+)
